Renamed company to business, added .env file section to application.yml, pom needs correcting (dependencies/plugins)
