+++ src/main/java/com/sam/business/ecom/translator/CustomerTranslator.java
@@ -1,8 +1,8 @@
-package com.sam.company.ecom.translator;
+package com.sam.business.ecom.translator;
 
-import com.sam.company.ecom.dto.CustomerRequestDTO;
-import com.sam.company.ecom.dto.CustomerResponseDTO;
-import com.sam.company.ecom.model.Customer;
+import com.sam.business.ecom.dto.CustomerRequestDTO;
+import com.sam.business.ecom.dto.CustomerResponseDTO;
+import com.sam.business.ecom.model.Customer;
 import org.springframework.stereotype.Component;
 
 @Component
