+++ src/main/resources/application.yml
@@ -1,4 +1,6 @@
 spring:
+  config:
+    import: optional:file:.env[.properties]
   datasource:
     username: ${DB_USER}
     url: ${DB_URL}
